add commit

+++ src/views/login/login body/LoginBody.js
@@ -21,32 +21,36 @@ const LoginBody = () => {
 
     const {email, password} = formData.current;
 
-    const data = {
-      email: email.value,
-      password: password.value
+    if (email.value === "" || password.value === "") {
+      alert("Missing form input parameters");
+    }
+    else {
+      const data = {
+        email: email.value,
+        password: password.value
+      }
+      axios.post('http://localhost:4000/api/login', data)
+      .then((response) => {
+        console.log(response)
+        const data = response.data;
+        const token = data.token;
+        if (!token) {
+            alert('Unable to login. Please try after some time.');
+            return;
+        }
+        localStorage.clear();
+        localStorage.setItem('user-token', token);
+        setTimeout(() => {
+          navigate('/dashboard',{state:{userInfo: data.user}})
+        }, 500);
+       
+      })
+      .catch(function (error) {
+        console.log(error);
+      });
     }
 
-    console.log(data)
 
-    axios.post('http://localhost:4000/api/login', data)
-    .then((response) => {
-      console.log(response)
-      const data = response.data;
-      const token = data.token;
-      if (!token) {
-          alert('Unable to login. Please try after some time.');
-          return;
-      }
-      localStorage.clear();
-      localStorage.setItem('user-token', token);
-      setTimeout(() => {
-        navigate('/dashboard')
-      }, 500);
-     
-    })
-    .catch(function (error) {
-      console.log(error);
-    });
  }
   return (
     <div className="login-div" >
@@ -54,24 +58,11 @@ const LoginBody = () => {
         <div className="login-form">
           <form ref={formData}>
             <p className="login-text">Login To Your Account</p>
-            <div className="login-icons">
-              <span className="facebook">
-                <img src={facebook} alt="facebook logo" />
-                <p className="icon-text">FACEBOOK</p>
-              </span>
-              <span className="twitter">
-                <img src={twitter} alt="twitter logo" />
-                <p className="icon-text">TWITTER</p>
-              </span>
-              <span className="google">
-                <img src={google} alt="google logo" />
-                <p className="icon-text">GOOGLE</p>
-              </span>
-            </div>
+            
 
             <div className="email-div">
               <hr />
-              <p>Or login with email</p>
+              <p>login with email</p>
               <hr />
             </div>
 
@@ -81,7 +72,7 @@ const LoginBody = () => {
                 id="email"
                 name="email"
                 className="name"
-                placeholder="Name"
+                placeholder="Email"
               />
               <input
                 type="password"
@@ -92,14 +83,10 @@ const LoginBody = () => {
             </div>
 
             <div className="checkbox-div">
-              <span>
-                <input type="checkbox" id="checkbox" className="checkbox" />
-                <p>Remember</p>
-              </span>
               <Link to="/forgotpassword">{<p className="forgot-password">Forgot Password?</p>}</Link>
             </div>
 
-            <button type="submit" onClick={onSubmit} className="continue">CONTINUE</button>
+            <button type="submit" onClick={onSubmit} className="continue">LOGIN</button>
           </form>
         </div>
